test the new status hint
